Fix CI failures: resolve 5 pyright errors and ruff formatting issues
- data_processing.py: Convert Path to str for tb.open_file() call
- datagen.py: Add type: ignore for control.matlab attr (missing stubs)
- infodynamics.py: Add type: ignore for JArray callable (jpype stubs)
- noderank.py: Cast gainmatrix mean to float for correct return type
- Fix ruff formatting in data_processing.py and weightcalc.py

+++ faultmap/noderank.py
@@ -482,7 +482,7 @@ def gainmatrix_preprocessing(gainmatrix: NDArray) -> tuple[NDArray, float]:
         gainsum += gainmatrix[col, row]
         counter += 1
 
-    currentmean = gainsum / counter
+    currentmean = float(gainsum / counter)
     meanscale = 1.0 / currentmean
 
     # Write meandiff to all gainmatrix elements indicated by connectionmatrix
